backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
azure-search-documents
openai
aiohttp
requests
google-adk
python-dotenv
python-multipart
google-api-python-client 
asyncpg
pandas
bcrypt==4.0.1 
passlib[bcrypt]==1.7.4
python-jose[cryptography]
pydantic[email]
scikit-learn
pandas
numpy
twilio
qdrant-client 
sentence-transformers
openpyxl
PyPDF2
apscheduler
google-generativeai
duckduckgo-search
